backend/requirements.txt
amqp==5.0.6
anyio==3.3.4
appnope==0.1.2
arabic-reshaper==2.1.3
asgiref==3.4.1
autopep8==1.5.7
backcall==0.2.0
billiard==3.6.4.0
black==21.12b0
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dadata==21.10.1
decorator==5.1.0
Django==3.2.8
django-cleanup==6.0.0
django-cors-headers==3.10.0
django-resized==0.3.11
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docopt==0.6.2
future==0.18.2
h11==0.12.0
html5lib==1.1
httpcore==0.14.3
httpx==0.21.1
idna==3.3
ipython==7.28.0
jedi==0.18.0
kombu==5.1.0
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
num2words==0.5.10
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pip-autoremove==0.10.0
platformdirs==2.4.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycodestyle==2.8.0
Pygments==2.10.0
PyJWT==2.2.0
pyotp==2.6.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-environ==0.4.54
pytz==2021.3
redis==3.5.3
reportlab==3.6.2
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
traitlets==5.1.0
typing_extensions==4.0.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.5
